[DEV][0007] | align product publish checkout state

+++ packages/product-command/db/queries/product_variants.sql
@@ -17,7 +17,7 @@ SELECT
     @product_id::uuid,
     @shop_id::uuid,
     unnest(@sku_codes::text[]),
-    unnest(@prices::int[]),
+    unnest(@prices::bigint[]),
     unnest(@currencies::text[]),
     unnest(@image_urls::text[]),
     unnest(@status::text[]),
